requirements.txt
RPi.GPIO
pygame
numpy